requirements.txt
Flask
python-dotenv
openai
flask-cors
requests
